feat(job offer): restrict job offers to a subset of the machines
Adds ability when creating the job offer to specify a list of machines that should
consider the offer. Machines that are not in the list will ignore the offer.
Engine keeps a picture of what's going on in the cluster.
So far it counts how many jobs each machine has.
Job offers are given to machines with low job count.