feat(job offer): restrict job offers to a subset of the machines
Engine keeps a picture of what's going on in the cluster.
So far it counts how many jobs each machine has.
Job offers are given to machines with low job count.
This is not a particular smart strategy, but it allowed me to put in place
the plumbing for more intelligence later.