fix(agent): Simplify purging jobs
The Agent.StopJob function is intended to be triggered by users and has
a very destructive impact on the cluster. When calling Agent.Purge, the
churn created by StopJob causes jobs that need to be rescheduled to
completely stop. Manually purging state and ignoring peers allows jobs
to reliably get rescheduled.