refactor(payload): Interact with Jobs, not Payloads
Payloads are not separate entities, they belong to Jobs. Calling code
should not have to explicitly create payloads in the registry.

This breaks the `fleetctl submit` behavior, as creation of the Job
earlier in the workflow is triggering it to get scheduled and
automatically started. This will only be broken temporarily.