fix(fleetctl): update ssh.Execute to use a PTY
This allows `fleetctl ssh` to work as expected with interactive commands
(like `toolbox`, or even just `bash`).
This change also reworks the error handling to ensure the exit code of
commands run over SSH are correctly propagated back to the client.
Finally, the semantics for copying input/output streams to/from sessions
are cleaned up (IMHO).