refactor(registry, tests): make keyPrefix configurable
Previously all registry access was hardcoded to a specific prefix
(/_coreos.com/fleet). This makes the endpoint configurable through both
fleet and the client. Accordingly, the functional tests are updated to
take advantage of this, which means they should no longer conflict with
any instances of fleet which happen to be running in the environment
used for functional testing.