refactor(registry): marshal only Name and UnitHash for jobs
Previously we were inadvertently encoding the entire Job in the Registry
when we really just wanted to store what's in jobModel - i.e. the Name
and UnitHash. This change moves to using jobModel for serialization as
well as deserialization.  Similarly, a jobOfferModel is used to
de/serialize JobOffers in the registry. Finally, errors encountered in
storing JobOffers are now propagated.