feat(fleetctl): add version check to fleetctl
This adds a basic version check/gate to fleetctl wherein it will print a
scary warning to the user if the version of fleetctl appears to be older
than the version of fleet running in the cluster. The version of fleet
in the cluster is inferred from the latest version found in all
machinestates registered in the registry.