fix(server): be more careful about shutting down
Server start-up has become a relatively complex process and there are
now various opportunities for it to block before the signal handlers
were set up appropriately. This makes it awkward to cleanly shut down
fleet. This change registers the signal handlers early in the process
and makes several changes to the initialization to facilitate this.