client/http: add new Unit methods
This builds on #756 and introduces parity in the HTTP client's
methods with those introduced to the registry client.

The current implementation is very dumb in that it simply wraps Jobs()
for all three methods.

This also exposed a small bug in the client in that the target state was
not being set correctly.