functional: simplify nspawn machine creation
Generating all SSH host keys every time we spin up a machine is
slow and unnecessary. We could consider moving to fixtures, but for now
it's easier to simply generate a single key.

systemd also introduced a hardware database service that was taking a
relatively considerable amount of time, and should be unecessary in a
systemd-nspawn container.

This brings the average machine bring-up to ~1s