agent: move AgentState entirely from Jobs to Units
As the title implies - AgentState (used by the agent reconciler for
desiredAgentState and also by the engine for clusterState) now uses
job.Unit instead of job.Job.

To facilitate this in the short term, Unit gains some of Job's methods
and simply wraps a Job to provide them. This will disappear as Job is
removed.