fleetd: fix support for "fleetd version"
Commit 32a51814 broke fleetd, making "fleetd version" the *only* thing it
could do and causing a index error panic otherwise. Now in addition to
properly checking the number of unparsed args exit with a usage error if
given any other unknown arguments.

Simply exiting if there are any unparsed args would have been sufficient
to fix #1210 but on the other hand allowing "fleetd version" doesn't
hurt so might as well keep it.

Fixes #1214 and #1210