Implement disable-engine flag
If set this fleetd will never assume mastership in a fleet cluster.
This will cut down on the interactions with etcd.

Useful when when you have two type of nodes in our clusters: worker and
admin nodes. The worker nodes come and go (in large numbers) and these
should just be used to schedule jobs. We never want them to become
leaders. The admin nodes control the entire cluster and we just want one
of those to become the fleet leader.

Add documentation and detail the consequences this can have.