fleetctl: Give DEBUG print if need wait for launch when start unit
When start unit, if [X-Fleet] give peer by "MachineOf=", the
fleetctl cannot know the exact reason. Only fleetd check the peer.
But still we can give a DEBUG print to inform launch is happing.
It's not the problem that the peer is not existing. After the
peer is running, the fleetctl will go on to the end. So just
add a debug print if --debug=true is given to fleetctl.

Fixes #1146