fleetctl: wait for unit creation to prevent spurious 404 errors
in testing with coretest, fleetd would get spurious errors while
creating and then reading back units. it is a race that occurs when a
PUT (consistent write to etcd) is immediately followed by a GET
(inconsistent read) to the same unit.

the real fix for this is to make use of etcd indexes everywhere in
fleetd, but unfortunately etcd 0.4.x indexes are broken. as a
workaround, we block in fleetctl waiting for acknowledgement of unit
creation before proceeding to load/start/etc.