fleetd: fix argument handling
We were calling userset.Usage but it was never actually defined. Unlike
in the top-level of the flag package (i.e. `flag.Usage`), this method
does not have a default value on `FlagSet`s.

The only reason this never showed up way earlier in the `ErrHelp` path
(and was only revealed via 8ebb81bcd1668be4a874db98cbf2e67ae305d7f6)
is that branch was never actually encountered.