fleetd: check for accidental `fleetctl` invocations
It's a common mistake for users to accidentally invoke fleet (which is,
unfortunately, a symlink to fleetd on CoreOS) instead of fleetctl, which
at best is confusing and at worst can cause problems if fleetd is
accidentally started up somewhere it shouldn't be and starts doing work
undesirably.

This is going to be backwards incompatible if anyone happened to be
invoking fleetd with arguments, but arguably in that case they were
doing it wrong anyway (since fleetd would just ignore them) and so this
provides better behaviour.