fleetctl:test: improve the start test and make sure that it handles templates
Since we create units from a template unit, there is a real wait
operation. To avoid that and be able to really test functions, just make
template units global by default, later new units will be global and we
won't wait forever to reach the targetstate we just assume the
Desiredstate.