functional: Add test for behaviour on etcd connectivity loss
This tests that no units are stopped when the agent loses the
connection, and that no superflous changes take place when the
connection is restored.

There are no checks for the behaviour of the rest of the cluster when
losing the node, nor for reconciliation after reconnect of the node to
the rest of the cluster.

(We use a one-node cluster for the test to make sure nothing goes on in
this regard.)