fleetd: keep a list of listeners across reconfiguring fleetd
So far fleetd has lost its existing listeners after receiving SIGHUP.
That happened because reconfigure() simply killed the server and started
a fresh one, without taking consideration into the existing listeners.
That resulted in a strange error like:

  Failed serving HTTP on listener: @ accept unix @: accept: invalid argument

So let's keep the list of existing listeners before killing the server,
and give it to the new server.

Fixes: https://github.com/coreos/fleet/issues/1190