Functional tests: Fix race condition in TestScheduleOneWayConflict()
Wait for the `destroy` command to take effect using a proper wait
condition (with `WaitForState()`), instead of a fixed sleep period --
which was sometimes insufficient, causing intermittent failure.

This fixes https://github.com/coreos/fleet/issues/1504

The condition function is mostly based on `WaitForNActiveUnits()`, with
some adjustments for the different scenario.