tests: fix TestParseFilepath where it is run under a unknown user or $HOME dir
Make sure that ParseFilepath() returned the same input since it was not
able to determin the current user or the $HOME env variable.

This fixes a bug where tests are under these conditions but at the same
time it adds unit checks for ParseFilepath() to make sure that it is
doing the right thing.