functional: introduce a new test TestDetectMachineId
A new test TestDetectMachineId checks if a etcd registration fails
when a duplicated entry for /etc/machine-id gets registered to
different machines. Note that it's expected to fail in this case.

Goal of the test is to cover the improvement patch by @wuqixuan
("fleetd: Detecting the existing machine-id").

See also https://github.com/coreos/fleet/pull/1288,
https://github.com/coreos/fleet/issues/1241,
https://github.com/coreos/fleet/issues/615.

Suggested-by: Olaf Buddenhagen <olaf@endocode.com>
Cc: wuqixuan <wuqixuan@huawei.com>