fleetd: introduce Replaces option in unit files
Introduce a new option 'Replaces' in X-Fleet the section, to get an
existing unit rescheduled to another machine. This option basically
works like an existing option 'Conflicts', except that circular replaces
in 2 units are not allowed. Also Conflicts cannot be used together with
Replaces, global cannot be used together with Replaces.

Fixes: https://github.com/coreos/fleet/issues/1394