agent: check errors from start/stop-ing unit
Now that UnitManager.Trigger{Start,Stop}() could return an error, the
error should be delivered up to mapTaskToFunc(), not to be ignored by
the task manager. Doing that, we can fix a bug in fleet ignoring failed
attempt to start/stop a unit.

Fixes https://github.com/coreos/fleet/issues/998