registry/rpc: skip TestRegistryMuxUnitManagement for travis CI
Skip TestRegistryMuxUnitManagement in case of an error from
NewSystemdUnitManager() to avoid error on travis CI.

  registrymux_test.go:99: unexpected error creating systemd unit manager:
  dial unix /var/run/dbus/system_bus_socket: connect: connection refused

Ideally we should fail with t.Fatalf(), but then it would always fail
on travis CI, because apparently systemd dbus socket is not available
there. So let's just skip the test for now. In the long run, we should
find a way to test it correctly on travis.