fleetctl: introduce option --max-print-units
For users who want to limit number of units to be printed, introduce an
option "--max-print-units" for these commands:
destroy, unload, stop, list-unit-files, and list-units.

Note that the default value of --max-print-units should be 0, which
means "all units are printed", so that the default behavior should
remain as before.

Fixes https://github.com/coreos/fleet/issues/1593