functional: allow WaitForNUnits to accept optional parameters
Allow Cluster.WaitForN{UnitFiles,Units,ActiveUnits} to accept opts, an
optional string that contains parameters to be passed to fleetctl.
As it's supposed to be optional, we don't need to modify call sites
of those functions: no opts for WaitFor*() means no opts for
nc.Fleetctl().

It's a preparation for a new test that is supposed to pass
"--max-print-units" option to the fleetctl command.