systemd: avoid iteration into unit list using UnitState
To avoid an additional iteration of unit states for each unit, set
UnitState.ActiveEnterTimestamp accordingly, by calling
m.systemd.GetUnitProperty(). Note that we need to retrieve the property
not only in the normal loop for the current systemd (>=230), but also
in the fallback handling for the older version of systemd (<=229).