glide: update vendor 2016-08-25
Update vendor tree to be up-to-date as of 2016-08-25.

Note that this update does not update google.golang.org/grpc, which
should stay with 178b68e2 ("Merge pull request #556 from
bradfitz/testcleanup", 2016-02-17). That's because the current gRPC
logic in fleet requires ClientConn.State(), which is currently not
available in gRPC 1.0. It's not trivial to change the logic.