fleetd: introduce an option systemd_link_unit_force
So far systemd.LinkUnitFiles() has been called with the force option
turned on. That was one of the reasons of performance issues.

Let's introduce an option systemd_link_unit_force, false by default.
Users can set the option in fleetd.conf.
That option will be passed directly to systemd.LinkUnitFiles().