agent,engine: use tri-state job schedule actions
So far decide() returns a bool variable to determine whether it's
possible to schedule or not. Reschedule was not handled in a proper way.
So let's create a new tri-state for job.JobAction, i.e.
JobActionSchedule, JobActionUnschedule, and JobActionReschedule.
And return the tri-state variable instead of bool.

That way AbleToRun() can check return values from HasConflict() and
HasReplace() in a correct manner.