fleetctl: take experimentalAPI into account in getClient
getClient() has not taken into account the case of !experimentalAPI,
before calling registryClient. It has set endPoint to a specific value
of URLs, but the new value would not be passed to getRegistryClient(),
which simply fetches endPoint again. This was a regression since
848d3561 ("fleetctl: convert cli to cobra").

To fix that, introduce getEndpoint() that does GetString("endpoint")
as well as the special handling for experimentalAPI. And make
getRegistryClient() call getEndpoint().

Fortunately, the experimentalAPI flag is set to true by default, so
this special handling path is not actively used after all. Just for
correctness.