fleetctl: make use of waitForState also in assertUnitState()
waitForUnitState() has been making use of an additional sleep, in case
assertUnitState() returned error. Now that a helper waitForState() is
available for fleetctl, we don't need to sleep there.

So remove the sleep calls, and instead make assertUnitState() a wrapper
of waitForState() over fetchUnitState(). That way, assertUnitState()
can periodically retry to get a unit, up to defaultSleepTime.
This is also good for consistency in the entire code, as
assertSystemdActiveState() already depends on waitForState().