functional: reduce total running time of all tests
As go test natively has an upper limit of 600sec in total for all
tests, we should reduce total running time.
So let's reduce number of units to be tested in the unit state tests,
and reduce maximum block attempts for the metadata template test.

This will fix the failure on semaphoreci, which has suddenly appeared
since the dynamic metadata tests. This should make tests work for now.
In the long run, it would be necessary to be able to tune the timeout.