functional: reduce number of tests in total
Functional test seems to have a limitation of 40 tests in total.
When the number of tests becomes more than 40, no more cluster
member can be created, so that test ends up with hanging mysteriously.

To avoid that, let's reduce number of tests, mainly for dynamic metadata
tests as well as unit action tests. It's just squashing similar tests.