functional: reduce total running time of all tests
As go test natively has an upper limit of 600sec in total for all
tests, we should reduce total running time.
So let's reduce number of units to be tested in the unit state tests,
and reduce maximum block attempts for the metadata template test.

Note that this patch was originally added via commit f81c3b2e, and
was reverted afterwards via b62c4c95, as a part of
https://github.com/coreos/fleet/pull/1709. However, now that the maximum
default sleep time of fleetctl got increased from 2 to 4 sec, we need
to reduce the total runtime, in order to avoid crashing of functional
tests. Without this change, the total test duration could end up with
again 600 seconds or longer.